minor fixes
git-svn-id: https://mspsim.svn.sourceforge.net/svnroot/mspsim/mspsim@495 23d1a52b-0c3c-0410-b72d-8f29ab48fe35

+++ se/sics/mspsim/net/IPv6Packet.java
@@ -52,8 +52,8 @@ public class IPv6Packet extends AbstractPacket {
   int version;
   int trafficClass;
   int flowLabel;
-  byte nextHeader;
-  byte hopLimit;
+  int nextHeader;
+  int hopLimit;
   long sourceAddressHi;
   long sourceAddressLo;
   long destAddressHi;
